move start.ts to main_process folder
